requirements.txt
arrow==1.4.0
asgiref==3.11.1
binaryornot==0.6.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cookiecutter==2.7.1
dj-database-url==3.1.2
Django==6.0.3
django-appconf==1.2.0
django-imagekit==6.1.0
django-tailwind==4.4.2
gunicorn==25.1.0
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
packaging==26.0
pilkit==3.0
pillow==12.1.1
psycopg2-binary==2.9.11
Pygments==2.19.2
pytailwindcss==0.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
PyYAML==6.0.3
requests==2.32.5
rich==14.3.3
six==1.17.0
sqlparse==0.5.5
text-unidecode==1.3
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.12.0
